ERGO/server/requirements.txt
adal==1.2.7
annotated-types==0.7.0
anyio==4.8.0
azure-core==1.32.0
azure-identity==1.20.0
azure-storage-blob==12.24.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
et_xmlfile==2.0.0
fastapi==0.115.8
greenlet==3.1.1
h11==0.14.0
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
numpy==2.2.3
openpyxl==3.1.5
pandas==2.2.3
portalocker==2.10.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
pywin32==308; sys_platform == "win32"
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
